+++ main.tf
@@ -6,18 +6,24 @@ resource "aws_instance" "rundeck" {
   instance_type          = var.instance_type
   vpc_security_group_ids = [aws_security_group.rundeck.id]
   subnet_id              = var.aws_subnet_id
-  tags                   = local.common_tags
-  user_data              = file("${path.module}/user_data.sh")
+  tags                   = var.tags
   iam_instance_profile   = local.use_instance_profile ? aws_iam_instance_profile.rundeck[0].name : null
+
+  user_data = templatefile("${path.module}/user_data.sh",
+    {
+      rdeck_jvm_settings = local.rdeck_jvm_settings == null ? "" : local.rdeck_jvm_settings
+    }
+  )
 
   root_block_device {
     volume_type           = "gp3"
     volume_size           = var.root_volume_size
     delete_on_termination = true
     encrypted             = var.root_encrypted
-    tags                  = local.common_tags
+    tags                  = var.tags
   }
 }
+
 
 # Spot Instance
 resource "aws_spot_instance_request" "rundeck" {
@@ -29,9 +35,14 @@ resource "aws_spot_instance_request" "rundeck" {
   wait_for_fulfillment           = true
   vpc_security_group_ids         = [aws_security_group.rundeck.id]
   subnet_id                      = var.aws_subnet_id
-  tags                           = local.common_tags
-  user_data                      = file("${path.module}/user_data.sh")
+  tags                           = var.tags
   iam_instance_profile           = local.use_instance_profile ? aws_iam_instance_profile.rundeck[0].name : null
+
+  user_data = templatefile("${path.module}/user_data.sh",
+    {
+      rdeck_jvm_settings = local.rdeck_jvm_settings == null ? "" : local.rdeck_jvm_settings
+    }
+  )
 
   ebs_block_device {
     device_name           = "/dev/sda1"
@@ -39,21 +50,14 @@ resource "aws_spot_instance_request" "rundeck" {
     volume_size           = var.root_volume_size
     delete_on_termination = true
     volume_type           = "gp3"
-    tags                  = local.common_tags
+    tags                  = var.tags
   }
 }
-
-# resource "aws_ec2_tag" "rundeck" {
-#   resource_id = aws_spot_instance_request.rundeck[0].spot_instance_id
-#   for_each    = local.common_tags
-#   key         = each.key
-#   value       = each.value
-# }
 
 # tag spot instance
 resource "aws_ec2_tag" "rundeck" {
   resource_id = var.create_spot_instance ? aws_spot_instance_request.rundeck[0].spot_instance_id : aws_instance.rundeck[0].id
-  for_each    = local.common_tags
+  for_each    = var.tags
   key         = each.key
   value       = each.value
 }
@@ -63,7 +67,7 @@ resource "aws_security_group" "rundeck" {
   name        = "rundeck-io-ec2"
   description = "Allow for Rundeck Servers"
   vpc_id      = var.aws_vpc_id
-  tags        = local.common_tags
+  tags        = var.tags
 
   ingress {
     from_port   = 22
@@ -94,14 +98,14 @@ resource "aws_iam_instance_profile" "rundeck" {
   count = local.use_instance_profile ? 1 : 0
   name  = "rundeck-io-instance-profile"
   role  = aws_iam_role.rundeck[count.index].name
-  tags  = local.common_tags
+  tags  = var.tags
 }
 
 # IAM role for rundeck host, only if ARNs passed into module
 resource "aws_iam_role" "rundeck" {
   count = local.use_instance_profile ? 1 : 0
   name  = "rundeck-io-role"
-  tags  = local.common_tags
+  tags  = var.tags
   assume_role_policy = jsonencode({
     Version = "2012-10-17",
     Statement = [