+++ variables.tf
@@ -86,3 +86,23 @@ variable "create_spot_instance" {
   default     = false
   description = "Create an EC2 Spot Instance"
 }
+
+variable "rdeck_jvm_settings" {
+  type        = string
+  default     = null
+  description = "Rundeck JVM Options"
+}
+
+variable "tags" {
+  type = map(string)
+  default = {
+    Name    = "Rundeck"
+    app     = "rundeck"
+    role    = "app"
+    ver     = "0.0.1"
+    contact = "hello@rundeck.io"
+    env     = "dev"
+    prov    = "terraform"
+  }
+  description = "Tags to use on AWS resources"
+}